Add `bufremove` plugin
With this plugin the functionality to close buffers is added to neovim.
The binding CTRL + W is added to close a buffer (in case the buffer was
edited by user then a confirmation panel will be displayed on the
screen). Also the binding CTRL + A is added to force close a buffer
(without asking for confirmation).